feat: Implement rubric-based scoring for job analysis
- Added categoryScoreSchema to define scoring categories with raw scores, weights, and contributions.
- Updated analysisResultSchema to include rubricVersion and categoryScores.
- Introduced a new scoring module (rubric.ts) to calculate fit scores based on weighted contributions from categories.
- Enhanced createOpenAIResponse function to include retry logic for API calls with logging for request metadata.
- Modified analyzeJobAndResumes function to integrate scoring logic and handle potential validation errors with controlled repairs.
- Updated database schema to change fit_score to numeric(4,1) for precise scoring and added ai_rubric_version to track rubric versions.
- Added comprehensive tests for schema validation, prompt structure, and error handling in careerCoach.test.ts.

+++ src/lib/validation.ts
@@ -91,7 +91,7 @@ export const jobFormSchema = jobBasicsSchema.extend({
     .optional(),
   // Registered with setValueAs (see AddJobDialog/JobDetailDialog) so an empty
   // input becomes null rather than coercing "" -> 0.
-  fitScore: z.number().int().min(0).max(10).nullable().optional(),
+  fitScore: z.number().min(0).max(10).nullable().optional(),
   resumeId: z.string().optional().or(z.literal("")),
   coverLetterUsed: z.string().trim().max(2000).optional().or(z.literal("")),
   priority: z.coerce.number().int().min(1).max(3).default(2),
